refactor: Navigationbar 비즈니스 로직 훅으로 분리
